Cleanup of tokenize.py and removal of obsolete modules.
Removed the line and manifest modules, switched the sample manifest
to the new manifest format.
Added comments throughout, better encapsulated functionality in
tokenize_stream(), fixed several typos in the state machine, and
made the enums more uniform.

Added nosetest support to the lib package.
In order to be scanned by nosetests, packages need to be identified as
such by an __init__.py file. Here we add these to the lib package and
subpackages.