Moved settings into lib, config loader cleanup
settings no longer will attempt to load a list of manifests specified
in the ini file, and will no longer search for a default ini file in
the /etc directory.
Added handling of sudo execution to settings by means of a copy of the
environment in which $USER is overwritten.