Added settings test module, rewrote settings imports
This adds the settings_test module with a basic test functions using
Mocks to generate a falsified version of the os module. This is the
first set of tests that relies on Mocks, but it must do so in order for
the tests to work independently of the existing local user accounts.

The os imports inside of the settings module were sometimes deeply
nested inside of functions and classes. This makes correctly mocking
them extremely difficult, which overall makes testing a pain. Now the
imports are at the outermost level, in the form of "import os", and
all of the mocking targets mocking specific attributes of the os
module.

Switched settings.py out of executable mode so that nosetests will run.