Separated tests from code, added tests requirements.txt
We used to keep all of the code and tests together in the same
directories. This makes actually reading through any body of code or
tests more difficult because you have an added burden of parsing
filenames for "_test" and correctly identifying the relevant elements.
In the interests of the long term health of the project, the tests are
being moved into a directory of their own, which in turn allows for
an easier separation of code from the testsuite in the future.