Expansion of settings tests, fixes for discovered bugs.
Improves handling for a set of active patches during a test.
Complete code coverage in settings under new tests.

Fixes a naming ambiguity in settings
Because a salve config object is itself configuration, the "conf"
attribute has unclear semantics. That dict has been renamed to
"attributes", and they are the attributes of the configuration -- as
distinct from the environment created by the configuration object.
Although the environment could be folded together with the attributes,
we do not do so to avoid potential naming conflicts, and to simplify
the string templating actions.