Renamed lib/parse to lib/reader.
Includes all of the necessary rewrites. This naming change is inspired
by the fact that lib/parse actually handles all of the manifest file
reading, not just the parsing of a token stream. There was a kind of
redundancy in the path spec lib.parse.parse . Hopefully, this will be
clearer.