Support for propogating config through blocks, moved salve.py out of src
This is a modification to ManifestBlocks, which need to support
expansion prior to runtime so that config errors and parsing errors are
apparent before any actions are run.
This includes a few new tests for ManifestBlocks that have been waiting
for this functionality.

Moved script to top level because it simplifies importing nonsense.