Added tests for util.streams and path expansion.
These are primarily tests that the hashing function works as desired.
The performance tests are intended to guard against unnecessary
work being introduced into the hashing function itself.

Added a missing test for file expand paths to increase coverage.
Should throw an error when the target is unspecified during path
expansion. This might be a bad place to check this in the future.