Modified commandline runs to print nicer messages.
Instead of throwing the exception, catches it, inspects it, and
prints a pretty error message. Requires some ugly manual handling of
all of the different exception types in case there is a hybrid exception
that inherits from multiple exception types. Looking for a more elegant
solution.

Includes testing that SALVEExceptions are caught, and other exceptions
pass through.