Fixes to DirBlock.to_action tests.
To accomodate the new structure of directory actions, many of the tests
had to be rewritten to acommodate action lists. Changes include handling
of mkdir when ugo.is_root returns true, and handling of recursive
copying of files (os.walk instead of 'cp -r').
A call to makedirs no longer takes place when os.path.exists returns
true on a desired directory.