Added execute.create actions to mkdirs and files
Instead of using touch and mkdir -p, we can use the os module and an
empty file write operation. This eliminates the last need for SAs in
the current implementation of SALVE (but we do not remove them because
they may be needed later).
Made the changes to use these in dir and file blocks.