Changed date format in backup log.
Altered from <year>-<month>-<day>.<epoch> to
<year>-<month>-<day> <hour>:<minute>:<second>
Both are at second-level resolution, but one is more human readable.

Includes addition of commandline backup-recovery argument that raises
an exception that indicates that SALVE does not yet have backup_recovery
support, in case it is invoked with that option before support is added.