Added action verification and verification codes.
Each Action has a set of verification codes which define the various failure
conditions of the Action, reasons that the Action was skipped, or the 'OK' status.
Verification codes are obtained from Action verification via Action.verify_can_exec()
verify_can_exec is called at the beginning of execution to obtain a verification code.
Depending on the verification code, warnings or errors may be written to stderr, and
the action may be skipped.
Checks are in place for copy actions. Cleanup for various elements of output production
in order to make warnings clear and concise.