Cleanup for CopyAction verify and added Create verification.
Verification will no longer produce an empty file when doing the writability check
for FileCopyActions, more testing takes place inside of a scratch dir for correctness,
DirCopy verification is now inspecting the source dir as well.
Includes modification of output to include the type of the warning based on the Action
class.