Added a verbosity setting and defaults for globals.
Although sensible defaults are specified through the default_settings.ini file,
these are exposed directly to the user for modification. It is safer, therefore,
for there to be a set of "true" defaults underlying these, which are set when
values are not specified.
Added min_verbosity constraint to INFO logging. Verbosity levels are adjustable
and modify which messages are seen.