Switched to OOP logging infrastructure
Logging is now done with a global Logger object, and the execution context it
works with is also global. Moves towards the possibility of exposing logging as
part of the SALVE API. Tests are failing due to I/O changes.
Began fixing the tests to mock the logger instead of trying to mock stderr &c.
By so doing, output control becomes much easier. As a mild requirement, means
that the logger must not be included with a `from` import in order for test
mocking to be simple.