Fix handling of strings for Python3
Switched use of string.letters to string.ascii_letters
Fixing logging with print(unicode(...)) requires some hackery of the "Are we
python 3?" variety. Basically, if unicode is not defined, we must be in
Python3, so redefine it as the desirable function.
Updates to hashing to make explicit calls to encode()