Moved salve.settings.config to salve.config
Rather than having a whole package devoted to settings, which does not properly
encapsulate the global execution context and other classifiable "settings", the
config package is exposed within the toplevel salve package. This is a clearer
breakdown. Config can be turned into a package if there is a call to do so in
the future.