Began work on abstraction layer for the filesys
The abstract files and directories are used as a kind of indirection to the
underlying filesys when normal operations are in progress. Their utility will
be apparent when operations are performed on a Filesystem Expectation that
tracks the future state of the filesys. When operating in this mode, it should
be possible to run the execute() methods of all actions without them touching
the underlying filesys.

While writing docstrings, did some fixup to Enum comments to use the commenting
style to match the salve.filesys docstrings in their use of the prompt '>>>'
when including python in a docstring.