Add initial implementation of a ConcreteFilesys
This is a Filesys which passes transparently through to the real, underlying
filesys. Should be a drop-in replacement for the calls to shutil and os
throughout the code. Add a failover constructor type to filesys lookups.
Testing required some minor adjustments to the FilesysElement class hierarchy
to be py2/3 compatible, embedding a small part of the 'six' library, provided
under the license and copyright of the six project. Extensive work on the
testing. Made CreateActions start using the Concrete FS.