Convert actions to take filesystem as an argument
Filesys is now an argument to action execution and verification. The root
action verification is done with the old real filesystem value from
salve.filesys, but in the future it can be preceded by explicit verification on
a virtual filesystem.
Actions now rely 100% on the filesys, not using 'os' directly for anything. Also
added a locations alias for 'os.path.join' to make this removal of 'os' easier to
curate.