Remove SHA512 speed test from testsuite
This has always been a questionable test. In fact, on the heavily loaded Travis
CI boxes, running pypy, this test fails. Apparently the pypy implementation of
sha512 is slower than the C implementation used in typical python. Not really
surprising, but it means that this test (and its implicit requirement) needs to
be dropped as we add pypy support.