Update setup.py to include default_settings.ini
This is part of the package_data field passed to the setup, as part of an
"include all .ini files" directive. This way, if more files are packaged up in
the future, they will be included as desired. Also includes argparse
dependency, fix to the license classifier, and the addition of the long
description (for presentation on pypi).