Bundled setup tooling into a Makefile
This is a little bit on the gross side, but the effect is that the Makefile can
be used to do a build without having to try to remember all of the build
tooling and commands. Overall a win. Default make target is the 'dev' build to
prevent an accidental promotion of a non-dev build into pypi when building