First draft version of v3 relative paths
Add an option to turn on SALVE version 3 style relative paths (relative to
current manifest, not root manifest). Additionally, adds a new general
"--version3" argument which enables all V3 options (of which, so far, there is
only one).
Includes a basic system test for running in --ver3 mode which expands the
system testing "run_on_manifest" to allow specification of commandline options,
and adds a simple test for file creation.