Stricter PEP8-ification of entire codebase
Newer verisons of PEP8 are more precise about enforcing under- and
over-indentation rules. These come into play especially with multiline strings,
and are present throughout the codebase. As I've recently upgraded the version
of PEP8 I run in my editor, it was sensible to try to update all of the code as
well.