Improve unit test coverage of dir blocks
Adds a needed unit test for dir block handling of recursive dir copy actions.
Revealed a minor "bug" in which paths were not normalized in CopyActions. The
results were not incorrect, but it would make logging harder read in an
automated fashion.
All dir block test are now wrapped in a ScratchContainer, which makes them more
uniform with other tests (this was not strictly necessary, but a nice-to-have)