Converted ExecutionContext to a Singleton
There's quite a bit of messiness associated with trying to pass around the
single ExecutionContext object and ensure that it is reachable from all parts
of the code. This shows up especially in the package __init__ files, and in
order to clean them up a sturdier solution was needed. To this end, the
ExecutionContext has been made into a Singleton using the Singleton metaclass
pattern, so it is accessed via class instantiation.
This complicates the testing somewhat -- the tests are now required to clear
the execution context during setup to ensure safety.