Switchover to Python stdlib 'logging'
There's still a global logger object, and it still behaves in much the same
way, so these changes are broad ranging, but small everywhere. The unnecessary
complexity of log types (INFO vs. DEBUG, etc.) plus verbosity levels is gone,
and "verbosity" as a numerical value will need to be replaced with something
that maps the count of '-v' flags onto sets of enabled logging channels.
The global logger is appropriate, since the logger is being used to print
errors and info about the run, not python exceptions and so forth. A logging
heirarchy would probably be overkill.
Because this is standard logging, not something custom, SALVE context info is
harder to handle. The ExecutionContext is autoprinted as part of a custom
formatter, but FileContext has to be stringified as part of the log message
when desired. Since there's a single global logger, it might make sense to
switch to some kind of wrapper/helper which takes that as an optional argument,
does string formatting, and then hands it off to the standard logger+formatter.
Such helpers would have to accept (*args,**kwargs) and pass through everything
other than the intercepted args[0] and filecontext=* to the logger, which makes
them slightly nontrivial.