Fix testsuite to run with proper logger
Ensures that logging output is suppressed/captured, and that it matches up
correctly with the expected values. Begins a general transition in tests to
check for error message inclusion (rather than exact output), as that better
represents the test "did we give the correct error?" without choking on
deprecations and other noise.
Also adds a few tests to get back to full test coverage.