Remove 'No handlers' line from log tests
In python2, an attempt to log when there are no handlers attached to a logger
makes noise on stderr. However, in python3, this is a silent no-op, and as a
result the check to see that the "No handlers could be found" message appears
will fail.
The easiest solution is to remove this check, since it's non-essential to the
test.