Special casing for Python 2.6 logging
The Python2.6 StreamHandler initializer has a different signature from the one
used in Python2.7 and Python3.x
Namely, rather than using a keyword arg, it uses an optional positional
argument for the Stream, and it defaults when the argument is omitted, not when
it is `None`. Although I'm unsure of the behavior of the logger on 2.6 when the
argument is supplied as None, I'd expect it to break. Per the documentation,
this is the only correct/safe way to set the stream for Python 2.6

In Python 2.6, the NullHandler was not part of the standard logging library. As
a result, we must wrap references to the NullHandler in salve.log.handler and
explicitly define a NullHandler when we get an AttributeError doing so (which
indicates that the logging module doesn't have an attribute named
'NullHanlder').

Fixes a typo of "'DEBUG'" instead of "logging.DEBUG"