Fix for config file overriding CLI logging option
Because config was being loaded after the CLI option processing, not before it,
the config file value of loglevel superseded the CLI option. This is the wrong
ordering, but to fix it, I had to move the config loading into a setting in the
execctx. Once that was done, usage sites for the config data change to lookups
in the execctx to find their config data.
Also broke some tests initially because they mock over the top of the wrong
config object. Updates these tests to (generally speaking) ignore the config
object handling and leave it to the behaviors of the test containment classes.