Add helper for checking action verification code
To try to simplify the action tests a little, add a helper that takes an
action and a verification code, checking that "verify_can_exec" will produce
that code.
In order for this to stay simple, the code is given as a str (and I added
getitem and setitem methods to the base Enum class).