Attempt to simplify tests with generators
There are a lot of near-identical tests in the testsuite. As an experiment to
see if they can be cleaned up by this strategy, convert the entire copy action
testsuite to be a set of test generators. They look a lot cleaner so far.
Also applied to file block tests and a few other locations.
Some other tests got some love from mock which they needed.