Convert Tokenizer to object
Tokenization is getting out of hand as a complex process with a lot of internal
state. Refactoring it into a Tokenizer object which encapsulates that state
simplifies a lot of it conceptually, since it removes the need for an
ExecutionContext key used to track this information. To accommodate this
change, broke the tokenizer into a subpackage: salve.parser.tokenize