Replace test generators with nose-parameterized
Generators for nose don't play nice with test classes because the generated
tests don't get proper setUp and tearDown function assignation. Clever hacks
with functools.wraps and similar don't seem to actually work, so a modification
to the overall testing framework is necessary.
There are big-honkin' parameter lists for the directory block tests, so these
aren't inlined with the parameterization decorator. This should be a model for
other locations -- define a global list of params to run a test over, and then
pass it to the decorator. In a sense, these parameter lists are tightly coupled
with the tests, but in another they're just "data" for the tests to consume.