Refactored action verification to be multi-phase
The structure of "verify_can_exec" being invoked as part of execute() was a
little funky, and it required that each action type do the same dance at the
beginning of execution. True, this is slated for removal, but that is a long
term goal, and in the meantime this has grown into a bit of a mess. Instead,
differentiates __call__() from execute() on actions by having __call__() do
verification before execution. Additionally, breaks down verification into two
steps: get a code (which tells us not only if we can proceed, but why or why
not), and then translate that code to a True/False value to tell us whether or
not to proceed. Without this breakdown, the 'OK' code has to act somewhat
truthy, which is not necessarily correct if we want other codes to allow
execution. For example, a "target exists" code could be fine for copy actions,
but still allows for more informative messaging than a raw "OK".